requirements.txt
authorizenet==1.0.14
Babel==2.5.1
bootstrap-admin==0.3.7.1
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-ads==0.1.8
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-classy-tags==0.8.0
django-countries==5.0
django-extra-views==0.6.4
django-haystack==2.7.dev0
django-localflavor==1.5.2
django-oscar==1.5
django-oscar-paypal==0.9.7
django-phonenumber-field==1.3.0
django-safedelete==0.4.3
django-sekizai==0.10.0
django-sslserver==0.20
django-tables2==1.14.2
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
django-wkhtmltopdf==3.1.0
factory-boy==2.9.2
Faker==0.8.6
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.18
lxml==4.1.1
mock==2.0.0
nose==1.3.7
olefile==0.44
pbr==3.1.1
phonenumberslite==8.8.5
Pillow==4.3.0
psycopg2==2.7.3.2
purl==1.3.1
pycountry==17.9.23
pysolr==3.6.0
python-dateutil==2.6.1
pytz==2017.3
PyXB==1.2.5
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4a1
text-unidecode==1.0
Unidecode==0.4.21
urllib3==1.22
